feat(addDot): make addDot a static function
