+++ src/app/modules/lobby/components/lobby-view/lobby-view.component.ts
@@ -20,6 +20,7 @@ import { IMultiplayerLobbySendFinalMessageDialogData } from 'app/interfaces/i-mu
 import { HttpErrorResponse } from '@angular/common/http';
 import { ToastType } from 'app/models/toast';
 import { ChallongeService } from 'app/services/challonge.service';
+import { Misc } from 'app/shared/misc';
 
 @Component({
 	selector: 'app-lobby-view',
@@ -349,17 +350,7 @@ export class LobbyViewComponent implements OnInit {
 	 * @param splitter the character to split the string with
 	 */
 	addDot(nStr: string | number, splitter: string) {
-		nStr = nStr.toString();
-		const x = nStr.split('.');
-		let x1: string = x[0];
-		const x2 = x.length > 1 ? `.${x[1]}` : '';
-		const rgx = /(\d+)(\d{3})/;
-
-		while (rgx.test(x1)) {
-			x1 = x1.replace(rgx, `$1${splitter}$2`);
-		}
-
-		return x1 + x2;
+		return Misc.addDot(nStr, splitter);
 	}
 
 	/**
